Separate out multi-component files
