+++ src/MLCollapse/MLCollapse.js
@@ -1,8 +1,6 @@
 import React from 'react'
-import PropTypes from 'prop-types'
 import { Collapse } from 'antd'
 import './style'
-const { Panel } = Collapse
 
 const MLCollapse = (props) => {
   return (
@@ -11,15 +9,5 @@ const MLCollapse = (props) => {
     </Collapse>
   )
 }
-
-const MLPanel = (props) => {
-  return (
-    <Panel {...props}>
-      {props.children}
-    </Panel>
-  )
-}
-
-MLCollapse.MLPanel = MLPanel
 
 export default MLCollapse
