+++ src/MLCollapse/MLPanel.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import { Collapse } from 'antd'
+const { Panel } = Collapse
+
+const MLPanel = (props) => {
+  return (
+    <Panel {...props}>
+      {props.children}
+    </Panel>
+  )
+}
+
+export default MLPanel
