+++ src/MLCollapse/index.js
@@ -1,2 +1,6 @@
 import MLCollapse from './MLCollapse'
+import MLPanel from './MLPanel'
+
+MLCollapse.MLPanel = MLPanel
+
 export default MLCollapse
