+++ src/MLDatePicker/index.js
@@ -1,2 +1,6 @@
 import MLDatePicker from './MLDatePicker'
+import MLRangePicker from './MLRangePicker'
+
+MLDatePicker.MLRangePicker = MLRangePicker
+
 export default MLDatePicker
