+++ src/MLLayout/index.js
@@ -1,2 +1,12 @@
 import MLLayout from './MLLayout'
+import MLHeader from './MLHeader'
+import MLContent from './MLContent'
+import MLSider from './MLSider'
+import MLFooter from './MLFooter'
+
+MLLayout.MLHeader = MLHeader
+MLLayout.MLContent = MLContent
+MLLayout.MLSider = MLSider
+MLLayout.MLFooter = MLFooter
+
 export default MLLayout
