+++ src/MLMentions/MLOption.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import { Mentions } from 'antd'
+const { Option } = Mentions
+
+const MLOption = (props) => {
+  return (
+    <Option {...props}>
+      {props.children}
+    </Option>
+  )
+}
+
+export default MLOption
