+++ src/MLMentions/index.js
@@ -1,2 +1,6 @@
 import MLMentions from './MLMentions'
+import MLOption from './MLOption'
+
+MLMentions.MLOption = MLOption
+
 export default MLMentions
