+++ src/MLTag/MLCheckableTag.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import { Tag } from 'antd'
+const { CheckableTag } = Tag
+
+const MLCheckableTag = (props) => {
+  return (
+    <CheckableTag {...props}>
+      {props.children}
+    </CheckableTag>
+  )
+}
+
+export default MLCheckableTag
