+++ src/MLTag/MLTag.js
@@ -1,8 +1,6 @@
 import React from 'react'
-import PropTypes from 'prop-types'
 import { Tag } from 'antd'
 import './style'
-const { CheckableTag } = Tag
 
 const MLTag = (props) => {
   return (
@@ -11,15 +9,5 @@ const MLTag = (props) => {
     </Tag>
   )
 }
-
-const MLCheckableTag = (props) => {
-  return (
-    <CheckableTag {...props}>
-      {props.children}
-    </CheckableTag>
-  )
-}
-
-MLTag.MLCheckableTag = MLCheckableTag
 
 export default MLTag
