+++ src/MLTag/index.js
@@ -1,2 +1,6 @@
 import MLTag from './MLTag'
+import MLCheckableTag from './MLCheckableTag'
+
+MLTag.MLCheckableTag = MLCheckableTag
+
 export default MLTag
